chore: move <PdfBox> into shared components folder
