+++ src/shared/components/pdf-box/PdfBox.css
@@ -25,7 +25,7 @@
             height: 100%;
             z-index: 1;
             opacity: 0.3;
-            background-image: url("../../../../../shared/media/backgrounds/background-pattern-1440p.png");
+            background-image: url("../../media/backgrounds/background-pattern-1440p.png");
             background-position: 40% 70%;
             background-size: 300%;
         }
